+++ movieDetail/view.html
@@ -122,8 +122,7 @@
       <p>Copyright © 2024 Node.js_4기 2주차 11조 All rights reserved.</p>
       <address>Contact webmaster for more information. https://github.com/ooommma/team11</address>
     </footer>
-    <script src="view.js"></script>
+    <script src="view.js" type="module"></script>
     <!-- IMPORT, EXPORT는 ES모듈6 이후 추가된 기능이기 때문에, type="module"을 명시해 줘야 작동함 -->
-    <script src="review.js" type="module"></script>
   </body>
 </html>
